pyproject.toml
[project]
name = "mcp-ai-tools-suite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "fastmcp>=2.10.5",
    "google-genai>=1.26.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
]
